updating AD for JAX.
